requirements.txt
Flask
line-bot-sdk
google-generativeai
python-dotenv
gunicorn
requests
